Adding 3 num to get to nearest target
This program states that there is a target given and from all the numbers we can achieve the nearest target.

It is an example of a 0-1 Knapsack Problem. The only difference is that there is no weight with the item only profit is used and the bag size is already specified.